More character validations, cleanup code
